bj/requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.9.0
